fixes in order to approve my pull request
