+++ src/main/java/org/fundacionjala/pivotal/pages/HeaderContainer.java
@@ -1,4 +1,26 @@
 package org.fundacionjala.pivotal.pages;
 
-public class HeaderContainer {
+import org.fundacionjala.core.ui.AbstractPage;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.springframework.stereotype.Component;
+
+/**
+ * this is the header container.
+ */
+@Component
+public class HeaderContainer extends AbstractPage {
+    /**
+     * this is the delete button that appears at the top of the page into a
+     * toast.
+     */
+    @FindBy(css = "button[title='Delete selected stories']")
+    private WebElement deleteStoryButtonOfToast;
+
+    /**
+     * this method clicks the delete button.
+     */
+    public void clickDeleteButtonOfToast() {
+        this.action.click(this.deleteStoryButtonOfToast);
+    }
 }
